Compare: make `CompareBytes` exported.
All the "magic" of reference comparison happens within the
`CompareBytes` function.

Reading a file (with `ioutil.ReadFile`) gives you an array of bytes and
gofpdf supports writing to a byte buffer, where we can easily get the
bytes from.

Exporting `CompareBytes` would make it possible to use this core
functionality instead of having to convert the data into `io.Reader`.