Added the ability to have multiple pages in a template (#216)
* Added the ability to have multiple pages in a template

Templates are now aware of the number of pages they have. The main
additions are the FromPage and FromPages methods which allow you to
extract a Template from a template if you pass in the page number. It
follows the same page mechanism as FPDF where pages start in the index
1.

* Update fpdf_test.go